+++ Portal/ServiceContainerCompilerPass/AddHttpMiddlewareCollectorCompilerPass.php
@@ -4,18 +4,16 @@ declare(strict_types=1);
 
 namespace Heptacom\HeptaConnect\Core\Portal\ServiceContainerCompilerPass;
 
-use Heptacom\HeptaConnect\Core\Web\Http\HttpMiddlewareClient;
-use Heptacom\HeptaConnect\Portal\Base\Web\Http\Contract\HttpClientMiddlewareInterface;
-use Psr\Http\Client\ClientInterface;
+use Heptacom\HeptaConnect\Core\Support\HttpMiddlewareCollector;
+use Psr\Http\Server\MiddlewareInterface;
 use Symfony\Component\DependencyInjection\Argument\TaggedIteratorArgument;
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Definition;
-use Symfony\Component\DependencyInjection\Reference;
 
-final class AddHttpMiddlewareClientCompilerPass implements CompilerPassInterface
+final class AddHttpMiddlewareCollectorCompilerPass implements CompilerPassInterface
 {
-    public const SERVICE_TAG = 'heptaconnect.http.client.middleware';
+    public const SERVICE_TAG = 'heptaconnect.http.handler.middleware';
 
     public function process(ContainerBuilder $container): void
     {
@@ -24,20 +22,18 @@ final class AddHttpMiddlewareClientCompilerPass implements CompilerPassInterface
         foreach ($definitions as $id => $definition) {
             $class = $definition->getClass() ?? (string) $id;
 
-            if (!\class_exists($class) || !\is_a($class, HttpClientMiddlewareInterface::class, true)) {
+            if (!\class_exists($class) || !\is_a($class, MiddlewareInterface::class, true)) {
                 continue;
             }
 
             $definition->addTag(self::SERVICE_TAG);
         }
 
-        $definition = (new Definition(HttpMiddlewareClient::class))
-            ->setDecoratedService(ClientInterface::class)
+        $definition = (new Definition(HttpMiddlewareCollector::class))
             ->setArguments([
-                new Reference(HttpMiddlewareClient::class . '.inner'),
                 new TaggedIteratorArgument(self::SERVICE_TAG),
             ]);
 
-        $container->setDefinition(HttpMiddlewareClient::class, $definition);
+        $container->setDefinition(HttpMiddlewareCollector::class, $definition);
     }
 }